split UI and sync signals
